added a mobile version of  header with JS functionality
