+++ src/html/template-parts/header/header.html
@@ -18,12 +18,12 @@
 </head>
 <body>
 	<div class="wrapper">
-		<header class="header">
+		<header class="header" id="body-lock">
 			<div class="container">
 				<div class="header-wrapper">
 					<a class="header-logo">
 						<p class="header-logo-text">
-							Сервисный центр <br> Apple в Москве
+							APP CENTER
 						</p>
 						<img src="img/header/logo.png" width="85" height="85" alt="">
 					</a>
@@ -36,15 +36,16 @@
 								<a href="">Акции</a>
 							</li>
 							<li class="header-submenu-link">
+								<a href="">Гарантии</a>
+							</li>
+							<li class="header-submenu-link">
 								<a href="">О компании</a>
 							</li>
 							<li class="header-submenu-link">
 								<a href="">Контакты</a>
 							</li>
 						</ul>
-						<div class="header-inner">
 						@@include( 'header-nav.html')
-						</div>
 						<p class="header-work">
 							Работаем с 9-00 до 22-00
 						</p>
