+++ src/js/common/global.js
@@ -1,0 +1,4 @@
+let targetElement
+
+export const getTargetElement = () => targetElement
+export const setTargetElement = element => targetElement = element
