+++ src/js/main.js
@@ -2,4 +2,4 @@ import './common/common'
 
 //burger-menu
 
-import './header/burger-menu'
+import './header/header'
